Reset skip error counter on a fabric link if it was down. (#3247)
What I did
reset the skip error counters for fabric link monitoring if a link was down and up again.

Why I did it

The fabric link monitoring feature does not take care the peer card restart cases. Peer card reload could lead to links go in to isolation state, and that's a case where we can ignore the error due to the init churn. So in this change , the skip error on init counters get reset if a link was down and up again.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>