[orchagent]: Skip installing ACL counter when ACL mirror rule is inactive (#3223)
* [orchagent]: Resolving issue #18844
* Don't install counters when session rule is inactive.
Mirror session is not activated and thus, the ACL rule is not created in sairedis/SAI (there is no next hop for the mirror destination IP).
However, orchagent creates and registers the ACL counter with the flexcounter (FC) and the counter is being polled every polling interval.
Since the ACL counter is not attached to any ACL rule, the BCM SAI introduces print errors in syslog.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>