Vlanmgrd handling of portchannel does not exist more gracefully. (#3367)
What I did:
Fixes sonic-net/sonic-buildimage#20684

Why I did:
It is seen when doing config reload on multi-asic (multiple instances of swss/teamd) their might be race condition when teammrd might be getting SIGTERM and cleaning up the port-channel in kernel and vlanmgrd might be adding the port channel as member in kernel and this can cause the exception as this operation as being done in async.

How I did:
To handle the failure gracefully if we get an exception when portchannel as vlan member in kernel we check if Portchannel exists and if not we just retry to minimize the risk of failure due to above race-condition.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>