Fix Voq chassis orchagent crash with 34K routes (#3329)
What I did
Don't add the remote system neighbor if the same neighbor exists.
Why I did it
The IMM has two asics and has 2 port channels in each asic and 2 port members in each port channel.
The ip address is configured on each port channel and bgp is enabled. The neighbor and routes are learned on these port channel.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>