[misc]: Add a dev folder for local development (#3368)
What I did
Add a dev folder, which includes a docker-compose environment for local development.

Why I did it
A developer has to download and install all dependences and setup the development environment if he would like to develop this submodule. This PR is for relieving the developer from the tedious setup process.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>