Fix the switch level Voq counter to use switch_id Oid instead of counter name for PACKET_INTEGRITY counter (#3521)
* Fix the switch level Voq counter to use switch_id Oid instead of counter name for PACKET_INTEGRITY counter
* Why I did it
The counter:oid table was using switch_id oid, but the COUNTERS_DEBUG_NAME_SWITCH_STAT_MAP was adding the counter name instead of the switch_id oid. So changed the COUNTERS_DEBUG_NAME_SWITCH_STAT_MAP to set the switch_id oid so that using the switch_id, the counters:oid: can be queried to get the counters for the packet_integrity counters.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>