[orchagent] Add value validation for RouteOrch, NhgOrch and Srv6Orch (#3596)
Add null checks for variable assignments in RouteOrch, NhgOrch and Srv6Orch. The assignments will only proceed if the string is not empty.

What I did
Add value validity checks to the assignment part of SET_COMMAND in the doTask methods for RouteOrch, NhgOrch and Srv6Orch, ensuring that the values assigned to fields are not empty.

Why I did it
For the value obtained by orchagent, assigning an empty value without validation may cause an issue.
So add a validity check before assigning the value.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>