[voq][chassis][orchagent][lagid] Modified the lua script to fix orhagent crash during upgrade from 202205 to 202405 (#3626)
* [voq][chassis][orchagent][lagid] Modified the lua script to fix orhagent crash duing upgrade from 202205 to 202405
What I did
Modify the lagids.lua script to check both SYSTEM_LAG_IDS_FREE_LIST and SYSTEM_LAG_ID_SET to make sure the LagId is really free before it is assigned. If it is in the FREE_LIST and also in the SYSTEM_LAG_ID_SET, remove it from the FREE_LIST and check/use the next one in the FREE_LIST. This will avoid using the ID which has been using by the LC which is still running the 202205 image. This will address the orchagent

Why I did it
The orhcgent crash during upgrade from 202205 to 202405 is caused by the same lagid used by multiple PortChannels. On chassis with mulitple LCs, upgrade LCs image from 202205 to 202405. orchagent crash could be occured when one LC is running with latest image while other LCs is still running with the old image. This PR fixes issue #https://github.com/Nokia-ION/ndk/issues/84

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>