Setting default nexthop weight to 1 in `fpmsyncd` (#3636)
* Set default nexthop weight to 1.
What I did

Nexthops are now considered to have a weight of 1 by default. This means that if libnl returns a weight of 0 for a nexthop, fpmsyncd will assume that its weight is 1.
Added a unit test for getNextHopWt to verify that it treats nexthops with a weight of 0 as having a weight of 1.
Every time a default weight is used for a nexthop, fpmsyncd will log an INFO-level message. The first time that this happens, a NOTICE-level message will also be logged.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>