Skip "port doesn't exist" SWSS_LOG_INFO messages for local ports (#3553)
* Don't emit "port does not exist" info logs for local ports
What I did
Modified the portsorch.cpp and neighorch.cpp to verify if a port shares the hostname with the local host. If so, we do not emit SWSS_INFO_LOG messages and these ports are eventually get erased from the consumer list.

Why I did it
Currently, we store local ports and remote ports using different convention in m_portList. Remote ports follow: || format and local ports are saved as . Hence when || format is used to 'get' a local port it returns false. This generates a number of "port not found" info messages in the syslog.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>