[muxorch] Update multi-mux routes for current neighbor update (#3691)
When a neighbor was updated, all routes with nexthops on the given mux port were being processed. In cases where a neighbor was deleted while a valid multi-mux route was still present and a different neighbor on the same mux port was updated, this could cause a crash (fixed in #3674).

Changes:

Change behavior such that only the current neighbor being updated is considered when searching for routes to be updated.

Add vstests to cover scenarios where different neighbors on a mux cable are added/removed while a valid multi-mux route is present.

What I did
Only process current multi-mux routes associated with current neighbor being updated

Why I did it
processing other routes caused crash when a nexthop associated with the route was deleted

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>