Enhance return code handling for SAI_SWITCH_ATTR_PACKET_TRIM_QUEUE_RESOLUTION_MODE (#3808)
What I did
Change the return code handling such that it does not produce an ERR level syslog in the case that SAI_STATUS_NOT_SUPPORTED is returned. Instead handle it gracefully.

Why I did it
Sonic seems to expect queryEnumCapabilitiesSai() to only be sucessful when it returns SAI_STATUS_SUCCESS. However, SAI_STATUS_NOT_SUPPORTED is also a valid return code if the attribute queried is not supported on a platform.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>