Skip flush if ring buffer is not empty or thread is not in idle (#3895)
* Skip flush if ring buffer is not empty or thread is not in idle
What I did
Skip flush in main thread if ring buffer is not empty or ring buffer thread is not in idle state.
This would avoid m_publisher to be triggered from both main thread and ring thread to cause race condition.

Why I did it
For fixing sonic-net/sonic-buildimage#24065

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>