[fpmsyncd]: Fix SRv6 SID List sharing among prefixes (#3860)
* [fpmsyncd]: Fix SRv6 SID List sharing among prefixes

Currently, there is a one-to-one mapping between prefixes and SRv6 SID
List. When fpmsyncd receives a new route, it creates a SID List and
associates it with the route. The key for the SID List is the Prefix of
route. If fpmsyncd receives a second route pointing to the same SRv6
SID List, fpmsyncd creates another copy of the SID List for the second
route.

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>