With ZMQ enabled between fpmsyncd to orchagent, default routes are se… (#3979)
* With ZMQ enabled between fpmsyncd to orchagent, default routes are set to DROP

Issue# https://github.com/sonic-net/sonic-swss/issues/3978

Recently ZMQ was enabled in sonic-mgmt (refer to sonic-net/sonic-mgmt@111e635).
fpmsyncd sends a DEL+SET for every route. This was coalesced by the producerStateTable infra. But with ZMQ enabled, this coalescing does not happen. As a result, orchagent gets a DEL+SET for every route. This works well normally. But for default routes, there is a bug in orchagent, where it adds a DROP action when it receives the DEL. But when the subsequent SET is received, it does not reset it to FORWARD action. This is due to a bug in its checking code.

* ZMQ configuration does not work when mgmt VRF is configured

Signed-off-by: Baorong Liu <96146196+baorliu@users.noreply.github.com>